parse the span the hard way

+++ src/letterboxd.rs
@@ -98,17 +98,17 @@ impl LetterboxdClient {
         }
     }
 
-    fn parse_rating2(rating: String) -> Result<u8> {
-        if rating.contains("rated-1") {return Ok(1 as u8);}
-        if rating.contains("rated-2") {return Ok(2 as u8);}
-        if rating.contains("rated-3") {return Ok(3 as u8);}
-        if rating.contains("rated-4") {return Ok(4 as u8);}
-        if rating.contains("rated-5") {return Ok(5 as u8);}
-        if rating.contains("rated-6") {return Ok(6 as u8);}
-        if rating.contains("rated-7") {return Ok(7 as u8);}
-        if rating.contains("rated-8") {return Ok(8 as u8);}
-        if rating.contains("rated-9") {return Ok(9 as u8);}
-        if rating.contains("rated-10") {return Ok(10 as u8);}
+    fn parse_rating2(rating: String) -> Result<usize> {
+        if rating.contains("rated-1") {return Ok(1 as usize);}
+        if rating.contains("rated-2") {return Ok(2 as usize);}
+        if rating.contains("rated-3") {return Ok(3 as usize);}
+        if rating.contains("rated-4") {return Ok(4 as usize);}
+        if rating.contains("rated-5") {return Ok(5 as usize);}
+        if rating.contains("rated-6") {return Ok(6 as usize);}
+        if rating.contains("rated-7") {return Ok(7 as usize);}
+        if rating.contains("rated-8") {return Ok(8 as usize);}
+        if rating.contains("rated-9") {return Ok(9 as usize);}
+        if rating.contains("rated-10") {return Ok(10 as usize);}
         return Err(anyhow!("unknown rating: '{}'", rating));
     }
